Related to G3D-4: set a different factor for layer and z-fighting (using w value); better matrix for perspective projection and z-fighting
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@65045 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/geogebra3D/euclidian3D/draw/DrawClippingCube3D.java
@@ -172,7 +172,7 @@ public class DrawClippingCube3D extends Drawable3DCurves {
 		double scaleMin = Math.min(Math.min(xscale, yscale), zscale);
 		int w = renderer.getWidth();
 		int h = renderer.getHeight();
-		int d = renderer.getVisibleDepth();
+		double d = renderer.getVisibleDepth();
 		frustumRadius = Math.sqrt(w * w + h * h + d * d) / (2 * scaleMin);
 
 		frustumInteriorRadius = Math.min(w, Math.min(h, d)) / (2 * scaleMax);
